requirements.txt
aiogram==2.25.1
aiohappyeyeballs==2.4.0
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.13.2
antlr4-python3-runtime==4.9.3
anyio==3.6.2
asteroid-filterbanks==0.4.0
async-timeout==4.0.2
attrs==23.1.0
audioread==3.0.1
Babel==2.9.1
certifi==2024.7.4
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.3
cmake==3.30.2
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.8.2
contourpy==1.2.1
cycler==0.12.1
decorator==4.4.2
dnspython==2.3.0
docopt==0.6.2
einops==0.8.0
emoji==2.2.0
face_recognition_models==0.3.0
fastapi==0.95.1
ffmpeg==1.4
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.3.3
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
huggingface-hub==0.24.6
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.8
imageio==2.35.1
imageio-ffmpeg==0.5.1
Jinja2==3.1.4
joblib==1.4.2
julius==0.2.7
kiwisolver==1.4.5
lazy_loader==0.4
librosa==0.10.2.post1
lightning==2.4.0
lightning-utilities==0.11.6
lit==18.1.8
llvmlite==0.43.0
magic-filter==1.0.9
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
more-itertools==10.4.0
motor==3.1.2
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.4
networkx==3.3
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
onnxruntime-gpu==1.19.0
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
opencv-python==4.10.0.84
optuna==3.6.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
primePy==1.3
proglog==0.1.10
protobuf==5.27.3
psutil==6.0.0
py-cpuinfo==9.0.0
pyannote.audio==3.3.1
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pycparser==2.22
pydantic==1.10.7
pydub==0.25.1
Pygments==2.18.0
pymongo==4.3.3
pyparsing==3.1.2
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.6
pytorch-lightning==2.4.0
pytorch-metric-learning==2.6.0
pytz==2024.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
rich==13.7.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scikit-learn==1.5.1
scipy==1.14.1
seaborn==0.13.2
semver==3.0.2
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.4.0
speechbrain==1.0.0
SQLAlchemy==2.0.32
srt==3.5.3
starlette==0.26.1
sympy==1.13.2
tabulate==0.9.0
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tiktoken==0.7.0
torch==2.4.0
torch-audiomentations==0.11.1
torch-pitch-shift==1.2.4
torchaudio==2.4.0
torchmetrics==1.4.1
torchvision==0.15.2
tqdm==4.66.5
triton==3.0.0
typer==0.12.4
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.81
ultralytics-thop==2.0.5
urllib3==2.2.2
uvicorn==0.21.1
vosk==0.3.45
websockets==13.0
whisper==1.1.10
yarl==1.9.2